python/pyproject.toml
[build-system]
requires = ["setuptools"]
backend = "setuptools.build_meta"

[project]
name = "insta-science"
authors = [
    {name = "John Sirois", email = "john.sirois@gmail.com"},
]
description = "A tool to help with installing and using science in Python projects."
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
dependencies = [
    "ansicolors",
    "appdirs",
    "filelock",
    "httpx",
    "packaging",
    "tomlkit; python_version < '3.11'",
    "tqdm"
]
dynamic = ["version"]

[project.scripts]
insta-science = "insta_science.shim:science"
insta-science-util = "insta_science.util:main"

[tool.setuptools.dynamic]
version = {attr = "insta_science.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["insta_science*"]

[dependency-groups]
dev = [
    "dev-cmd",
    "mypy",
    "pytest",
    "pytest-xdist",
    "ruff",
    "types-appdirs",
    "types-tqdm",
    "typing-extensions",
]

[[tool.mypy.overrides]]
module = ["colors.*"]
follow_untyped_imports = true

[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I"]

[tool.dev-cmd.commands]
clean = [
    "python",
    "-c",
    """\
import os
import shutil
import sys

import colors


# The MyPy cache can get corrupted by swithing Pythons with `uv ... --python ...` and re-building
# the .venv is cheap; so we clean these as a more restricted form than `git clean -fdx`
for cache_dir in '.mypy_cache', '.venv':
    if os.path.exists(cache_dir):
        shutil.rmtree(cache_dir, ignore_errors=True)
        print(colors.green(f"Removed directory `{cache_dir}`."), file=sys.stderr)
    """
]

fmt = ["ruff", "format"]
check-fmt = ["ruff", "format", "--diff"]

lint = ["ruff", "check", "--fix"]
check-lint = ["ruff", "check"]

type-check = ["mypy", "insta_science"]
type-check-39 = ["mypy", "--python-version", "3.9", "insta_science"]
type-check-313 = ["mypy", "--python-version", "3.13", "insta_science"]

[tool.dev-cmd.commands.test]
args = ["pytest", "-n", "auto"]
accepts-extra-args = true

[tool.dev-cmd.aliases]
checks = ["fmt", "lint", "type-check-39", "type-check-313", "test"]
ci = ["check-fmt", "check-lint", "type-check", "test"]

[tool.dev-cmd.default]
alias = "checks"
